Detach PopularModelsRatio.js and TopTokenUserList.js from DashboardPage.js
